Move Logging to impl package
