+++ src/commonMain/kotlin/io/klogging/impl/KloggerImpl.kt
@@ -20,7 +20,6 @@ package io.klogging.impl
 
 import io.klogging.Klogger
 import io.klogging.Level
-import io.klogging.Logging
 import io.klogging.context.LogContext
 import io.klogging.events.LogEvent
 import io.klogging.events.currentContext
